Add no implicit any and make parcel compile index.html
